move integrations + docs
